Refactoring - move Sd Card utilities out of AppContext into SdUtils.
