+++ app/src/main/java/com/door43/translationstudio/newui/home/HomeActivity.java
@@ -37,6 +37,7 @@ import com.door43.translationstudio.newui.newtranslation.NewTargetTranslationAct
 import com.door43.translationstudio.newui.FeedbackDialog;
 import com.door43.translationstudio.newui.translate.TargetTranslationActivity;
 import com.door43.translationstudio.AppContext;
+import com.door43.translationstudio.util.SdUtils;
 import com.door43.widget.ViewUtil;
 
 import org.apache.commons.io.FileUtils;
@@ -211,7 +212,7 @@ public class HomeActivity extends BaseActivity implements WelcomeFragment.OnCrea
     }
 
    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == AppContext.REQUEST_CODE_STORAGE_ACCESS) {
+        if (requestCode == SdUtils.REQUEST_CODE_STORAGE_ACCESS) {
             Uri treeUri = null;
             String msg = "";
             if (resultCode == Activity.RESULT_OK) {
@@ -219,7 +220,7 @@ public class HomeActivity extends BaseActivity implements WelcomeFragment.OnCrea
                 // Get Uri from Storage Access Framework.
                 treeUri = data.getData();
                 final int takeFlags = data.getFlags();
-                boolean success = AppContext.validateSdCardWriteAccess(treeUri, takeFlags);
+                boolean success = SdUtils.validateSdCardWriteAccess(treeUri, takeFlags);
                 if(!success) {
                     String template = getResources().getString(R.string.access_failed);
                     msg = String.format(template,treeUri.toString());
